implement the top layer window
As we show/clip the windows tree from left to right, from bottom to top,
an always-on-top window is the window that should always be in left.
It's tricky but we could get the work done by changing little of code.
The diff is pretty small.

Create window with RTGUI_WIN_STYLE_ONTOP to let the window be in top
layer.