suppress the compiling cmd line by default
The cmd line of compliling is lenghty and there is no need to see them
in most of the cases. Suppress them will make warnings more clearer to
see. However, SCons 2.1.0 has a bug on LINKCOMSTR on VC tools. So the
LINKCOMSTR doesn't work.

If you want to see the full cmd line output, use `scons.bat VERBOSE=1`.