let wintitle inherit from widget and remove toplevel
Actually the title is not a window at all. It could be a sub class of
widget and managed totally by the window manager(topwin).

As a consequence, the only class that inherit toplevel is win. So I just
move the toplevel into win and removed that class. There will be only
one "toplevel" class -- rtgui_win.