don't update window on EVENT_WIN_{ACTIVATE,DEACTIVATE}
This fix a double-paint bug which is originally found by onelife. Now
the paint event of window is managed by topwin. It is guarantee that
activate event will be followed by a paint event. So if you do painting
in on_activate and it blow up your system, that's your problem.