app: fix api, always use struct rtgui_app* as the first parameter
Firstly, I want to expose struct rtgui_app to the userspace. Just
because: 1) most of other parts of RTGUI did this; 2) the struct has
some information that userspace might be interested with. Although this
behavior is strongly volatile the rule of "information hiding", there
are so many information has been already leaked in RTGUI and such piece
of info does not count much.

Secondly, I want the "Object-Oriented" coding style apply to this piece
of code as well. So the first parameter of function should be the
instance of class -- struct rtgui_app*. This result in more consistence
in API appearance because some API like rtgui_app_{activate,close} will
always need struct rtgui_app*.