new font
Frankly speaking, I'm not sure which font is it. I generated it with
PCtoLCD using SongTi font. But the result is different with it. I have
installed CourierStd before generating the font. But the font looks
different from it too.

But it looks good in RTGUI. I think that's enough.