add calibration to RTGUI
This moves the calibration.c from BSPs into RTGUI. The only thing that
the touch driver should do is to #include <rtgui/calibration.h> to get
the struct calibration_data definition.

It provide calibration_set_restore/calibration_set_after to give ways to
the user to save and restore the calibration data.